+++ src/site/markdown/overview/changelog.md
@@ -4,9 +4,12 @@
 
 **Feature Requests and Improvements:**
 
+* [#1293](https://sourceforge.net/p/pmd/bugs/1293/): Disable VariableNamingConventions for native methods
+
 **New/Modified Rules:**
 
 * [Java / Design / UseVarargs](http://pmd.sourceforge.net/pmd-java/rules/java/design.html#UseVarargs): if `byte[]` is used as the last argument, it is ignored and no violation will be reported.
+* [Java / Naming / VariableNamingConventions](http://pmd.sourceforge.net/pmd-java/rules/java/naming.html#VariableNamingConventions): New property `checkNativeMethodParameters`
 
 **Pull requests:**